+++ src/com/nixinova/mineo/ui/graphics/ScreenText.java
@@ -3,7 +3,6 @@ package com.nixinova.mineo.ui.graphics;
 import java.awt.Color;
 import java.awt.Graphics;
 
-import com.nixinova.mineo.io.Options;
 import com.nixinova.mineo.main.Game;
 import com.nixinova.mineo.main.Mineo;
 import com.nixinova.mineo.ui.display.GameDisplay;
@@ -12,10 +11,6 @@ public class ScreenText {
 
 	private static final int LEFT_MARGIN = 5;
 	private static final int TOP_MARGIN = 5;
-	private static final int SEP = 15;
-
-	private final FontGraphics headerFont = FontGraphics.FONT_150;
-	private final FontGraphics infoFont = FontGraphics.FONT_100;
 
 	private Graphics graphics;
 	private int curLineIndex = 0;
@@ -25,6 +20,9 @@ public class ScreenText {
 	}
 
 	public void drawMainInfo(Game game) {
+		var headerFont = FontGraphics.FONT_150;
+		var infoFont = FontGraphics.FONT_100;
+
 		boolean showFullInfo = game.controls.gameInfoShown;
 
 		// Draw heading
@@ -48,36 +46,39 @@ public class ScreenText {
 		if (lookingAt != null)
 			drawInfoLine(infoFont, "Selected: %d / %d / %d", lookingAt.x, lookingAt.y, lookingAt.z);
 	}
+	
+	public void checkThenDrawSavedDataWarning(String type, float fileV, float curV) {
+		var warningFont = FontGraphics.FONT_200;
 
-	public void drawOptionsWarning() {
 		String msg1 = "", msg2 = "", msg3 = "";
-		float fileV = Options.fileVersion;
-		float curV = Options.OPTIONS_VERSION;
 		boolean diffMajor = (int) fileV != (int) curV;
 		boolean diffMinor = (int) (fileV * 10) != (int) (curV * 10);
 		if (diffMinor) { // only breaking change if major or minor is different
 			if (fileV < curV)
-				msg1 = "Outdated options version! Client is on " + curV + " while options.txt is on " + fileV + ".";
+				msg1 = String.format("Outdated %1$s version! Client is on %2$.2f while %1$s is on %3$.2f.", type, curV, fileV);
 			else
-				msg1 = "Options file too new! options.txt is on " + fileV + " while client is on " + curV + ".";
+				msg1 = String.format("%1$s file is too new! %2$.2f is newer than expected version %3$.2f.", type, curV, fileV);
 
 			if (diffMajor) // breaking changes to the file format
-				msg2 = "Data in options.txt which differs from the current version may break or crash your game!";
+				msg2 = String.format("Data in %s which differs from the current version may break or crash your game!", type);
 			else if (diffMinor) // changes to implementation of values
-				msg2 = "Data in options.txt which differs from the current version may not work correctly!";
+				msg2 = String.format("Data in %s which differs from the current version may not work correctly!", type);
 
-			msg3 = "Delete options.txt in %appdata%\\.mineo and restart the game to refresh the options file.";
+			msg3 = String.format("Update %s in %%appdata%%\\.mineo and restart the game to refresh the file.", type);
 		}
 
 		Color col = diffMajor ? Color.red : diffMinor ? Color.yellow : Color.white;
-		infoFont.drawString(graphics, msg1, col, LEFT_MARGIN, GameDisplay.HEIGHT - SEP * 8);
-		infoFont.drawString(graphics, msg2, col, LEFT_MARGIN, GameDisplay.HEIGHT - SEP * 7);
-		infoFont.drawString(graphics, msg3, col, LEFT_MARGIN, GameDisplay.HEIGHT - SEP * 6);
+		var textScheme = new TextColorScheme(col, Color.black);
+		int startX = GameDisplay.WIDTH / 5;
+		curLineIndex++;
+		warningFont.drawStringOutlined(graphics, msg1, startX, 20 * curLineIndex++, textScheme);
+		warningFont.drawStringOutlined(graphics, "  " + msg2, startX, 20 * curLineIndex++, textScheme);
+		warningFont.drawStringOutlined(graphics, "  " + msg3, startX, 20 * curLineIndex++, textScheme);
 	}
 
 	private void drawInfoLine(FontGraphics fg, String fStr, Object... args) {
 		String fmtdString = String.format(fStr, args);
-		fg.drawString(graphics, fmtdString, Color.white, LEFT_MARGIN, TOP_MARGIN + SEP * curLineIndex++);
+		fg.drawString(graphics, fmtdString, Color.white, LEFT_MARGIN, TOP_MARGIN + 15 * curLineIndex++);
 	}
 
 }